requirements.txt
google-cloud-aiplatform
pymupdf
requests
google-generativeai
google-genai
google-cloud-storage
pypdf